+++ conf/trigger_settings.py
@@ -261,8 +261,8 @@ AUTOACL_FILE = os.environ.get('AUTOACL_FILE', os.path.join(PREFIX, 'autoacl.py')
 # Loader's module, subsequent items are passed to the Loader during
 # initialization.
 NETDEVICES_LOADERS = (
-    'trigger.netdevices.loaders.filesystem.XMLLoader',
     'trigger.netdevices.loaders.filesystem.JSONLoader',
+    'trigger.netdevices.loaders.filesystem.XMLLoader',
     'trigger.netdevices.loaders.filesystem.SQLiteLoader',
     'trigger.netdevices.loaders.filesystem.CSVLoader',
     'trigger.netdevices.loaders.filesystem.RancidLoader',
@@ -274,7 +274,9 @@ NETDEVICES_LOADERS = (
 # A path or URL to netdevices device metadata source data, which is used to
 # populate trigger.netdevices.NetDevices. For more information on this, see
 # NETDEVICES_LOADERS.
-NETDEVICES_SOURCE = os.environ.get('NETDEVICES_SOURCE', os.path.join(PREFIX, 'netdevices.xml'))
+NETDEVICES_SOURCE = os.environ.get(
+    'NETDEVICES_SOURCE', os.path.join(PREFIX, 'netdevices.json')
+)
 
 # Assign NETDEVICES_SOURCE to NETDEVICES_FILE for backwards compatibility
 NETDEVICES_FILE = NETDEVICES_SOURCE